clippy

+++ src/hc14/setting/command.rs
@@ -112,7 +112,7 @@ impl GenerateAtCommand for Channel {
             }
             for i in buf.iter_mut() {
                 if n > 0 {
-                    *i = (n % 10) as u8 + b'0';
+                    *i = (n % 10) + b'0';
                     n /= 10;
                 } else {
                     *i = b"0"[0];
